requirements.txt
# --- Core Rasa ---
rasa==3.6.21
rasa-sdk==3.6.2

# --- ML / NLP ---
transformers==4.35.0
sentence-transformers==2.5.1
huggingface-hub==0.17.3
torch==2.1.0
scikit-learn==1.1.3

# --- Optional fast vector search (comment out if fails) ---
# faiss-cpu==1.7.4

# --- Runtime & utilities ---
uvicorn==0.20.0
numpy==1.23.5
typing-extensions==4.9.0
sqlalchemy<2.0
accelerate>=0.26.0
pandas>=1.5.0,<2.2.0
PyYAML==6.0.2
Flask==2.3.2
requests==2.31.0
fastapi==0.111.0
tqdm
python-dotenv
